generic-animation/requirements.txt
viam-sdk
pypubsub